requirements.txt
langchain
langchain_community
langchain_google_genai
python_dotenv
flask
flask-RESTful
faiss_cpu
fpdf
pypdf
pandas